moved talk_to_agent to server's standard tools
