+++ src/backend/a2a_client.py
@@ -109,6 +109,7 @@ class AgentBeatsA2AClient:
                         agent_name=agent_name
                     )
             
+            # kickoff signal
             battle_info = {
                 "type": "battle_start",
                 "battle_id": battle_id,
